Move code lens files into own directory
