+++ extensions/typescript-language-features/src/languageFeatures/codeLens/baseCodeLensProvider.ts
@@ -5,11 +5,11 @@
 
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
-import type * as Proto from '../protocol';
-import { ITypeScriptServiceClient } from '../typescriptService';
-import { escapeRegExp } from '../utils/regexp';
-import * as typeConverters from '../utils/typeConverters';
-import { CachedResponse } from '../tsServer/cachedResponse';
+import type * as Proto from '../../protocol';
+import { CachedResponse } from '../../tsServer/cachedResponse';
+import { ITypeScriptServiceClient } from '../../typescriptService';
+import { escapeRegExp } from '../../utils/regexp';
+import * as typeConverters from '../../utils/typeConverters';
 
 const localize = nls.loadMessageBundle();
